+++ core/src/main/java/module-info.java
@@ -3,5 +3,6 @@
  */
 module com.github.cowwoc.pouch.core
 {
+	requires org.slf4j;
 	exports com.github.cowwoc.pouch.core;
 }